automatically create enum based on return type of action creator impl
